api/requirements.txt
boto3==1.28.57
deepdiff==5.6.0
Flask==2.3.2
Flask_Cors==3.0.10
neo4j==5.7.0
bcrypt==3.2.2
cryptography==42.0.4
sklearn==0.0
pandas==1.5.3
python-dotenv==1.0.0
numpy==1.22.0
pydash==7.0.6
jsonschema>=4.18.0
gunicorn==20.1.0
watchdog==3.0.0
urllib3<2.0.0
